Initial Commit
This is my version of "Raytracer in a Weekend"

So far I am most of the way through the first book.  I need to add the
camera stuff, but that's mostly it.

I have added multithreading to speed everything up, although it's at the
cost of making this implementation Windows specific - at least until I
add some interface for threading so I can add pthreads support.